requirements.txt
torch
transformers
accelerate
peft
bitsandbytes
fastapi
uvicorn
gunicorn
pydantic-settings
pyyaml
scipy